Add types to delta service
